Refactor tests: Migrate redundant integration tests to domain unit tests
- Migrated backup target resolution tests from `tests/library/backup_target.rs` to unit tests in `src/domain/backup_target.rs`.
- Removed `tests/library/backup_target.rs` and cleaned up `tests/library/mod.rs`.
- Removed redundant `domain_profile_types_are_public` test from `tests/library/domain_exports.rs` as the underlying logic (`resolves_canonical_profiles`) is adequately covered in `src/domain/profile.rs` unit tests.

These changes centralize domain logic tests within the domain module, reducing redundant maintenance overhead and technical debt.
