+++ src/domain/backup_target.rs
@@ -59,3 +59,40 @@ impl fmt::Display for BackupTarget {
         f.write_str(self.name())
     }
 }
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn backup_target_resolves_system() {
+        let target = BackupTarget::from_input("system");
+        assert!(target.is_some());
+        assert_eq!(target.unwrap().name(), "system");
+    }
+
+    #[test]
+    fn backup_target_resolves_vscode() {
+        let target = BackupTarget::from_input("vscode");
+        assert!(target.is_some());
+        assert_eq!(target.unwrap().name(), "vscode");
+    }
+
+    #[test]
+    fn backup_target_resolves_vscode_extensions_alias() {
+        let target = BackupTarget::from_input("vscode-extensions");
+        assert!(target.is_some());
+        assert_eq!(target.unwrap().name(), "vscode");
+    }
+
+    #[test]
+    fn backup_target_rejects_unknown() {
+        assert!(BackupTarget::from_input("unknown").is_none());
+    }
+
+    #[test]
+    fn backup_target_all_returns_expected_set() {
+        let all = BackupTarget::all();
+        assert_eq!(all.len(), 2);
+    }
+}
