+++ tests/library/domain_exports.rs
@@ -1,13 +1,6 @@
 //! Verify public API surfaces remain accessible.
 
 #[test]
-fn domain_profile_types_are_public() {
-    // Confirms that profile resolution is accessible from the library crate.
-    let resolved = mev::domain::profile::resolve_profile("macbook");
-    assert_eq!(resolved, Some("macbook"));
-}
-
-#[test]
 fn domain_tag_resolution_is_public() {
     let tags = mev::domain::tag::resolve_tags("rust");
     assert_eq!(tags, vec!["rust-platform", "rust-tools"]);
