+++ tests/library/mod.rs
@@ -1,2 +1,1 @@
-mod backup_target;
 mod domain_exports;
